refactor(bank-account-schemas): consolidate USA and CAN schemas into one file and delete CAN file; align with core-info pattern
