--- packages/webcomponents/src/components/business-forms/schemas/business-bank-account-schema-canada.ts
@@ -1,80 +1,0 @@
-import { object, mixed, string } from 'yup';
-import { EntityDocumentType } from '../../../api/Document';
-import {
-  accountNumberValidation,
-  accountTypeValidation,
-  bankNameValidation,
-  identityNameValidation,
-  nicknameValidation,
-  numbersOnlyRegex,
-} from './schema-validations';
-
-const checkExistingDocs = (documents: any[], documentType: EntityDocumentType) => {
-  return documents.some((doc) => doc.document_type === documentType);
-}
-
-const documentErrorText = 'Please upload either a voided check or a bank statement. Only one is required.';
-
-export const businessBankAccountSchemaCanada = (documents: any[], allowOptionalFields?: boolean) => {
-  const existingVoidedCheck = checkExistingDocs(documents, EntityDocumentType.voidedCheck);
-  const existingBankStatement = checkExistingDocs(documents, EntityDocumentType.bankStatement);
-  const existingDocument = existingVoidedCheck || existingBankStatement;
-
-  const institutionNumberValidation = string()
-    .length(3, 'Institution number must be 3 digits')
-    .matches(numbersOnlyRegex, 'Enter valid institution number');
-
-  const transitNumberValidation = string()
-    .length(5, 'Transit number must be 5 digits')
-    .matches(numbersOnlyRegex, 'Enter valid transit number');
-
-  const schema = object().shape(
-    {
-      bank_name: bankNameValidation.required('Enter bank name'),
-      nickname: nicknameValidation.required('Enter nickname'),
-      account_owner_name: identityNameValidation.required('Enter account owner name'),
-      account_type: accountTypeValidation.required('Select account type'),
-      account_number: accountNumberValidation.required('Enter account number'),
-      institution_number: institutionNumberValidation.required('Enter institution number'),
-      transit_number: transitNumberValidation.required('Enter transit number'),
-      voided_check: mixed().when('bank_statement', {
-        is: (val: any) => {
-          return !val && !existingDocument;
-        },
-        then: (schema) => {
-          return schema.required(documentErrorText);
-        },
-        otherwise: (schema) => {
-          return schema.notRequired();
-        }}
-      ),
-      bank_statement: mixed().when('voided_check', {
-        is: (val: any) => {
-          return !val && !existingDocument;
-        },
-        then: (schema) => {
-          return schema.required(documentErrorText);
-        },
-        otherwise: (schema) => {
-          return schema.notRequired();
-        }}
-      ),
-    }, [['voided_check', 'bank_statement']]
-  );
-
-  const easySchema = object({
-    bank_name: bankNameValidation.nullable(),
-    nickname: nicknameValidation.nullable(),
-    account_owner_name: identityNameValidation.nullable(),
-    account_type: accountTypeValidation.nullable(),
-    account_number: accountNumberValidation.nullable(),
-    institution_number: string().nullable(),
-    transit_number: string().nullable(),
-    voided_check: mixed().nullable(),
-    bank_statement: mixed().nullable(),
-  });
-
-  return allowOptionalFields ? easySchema : schema;
-}
-
-
